chore: move more cells into its own file
